🧩 Y61ZHN task: split provider conflict modules
