+++ .agentplane/tasks/202607302125-Y61ZHN/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607302125-Y61ZHN/record-superseded-provider-conflict-outcomes-wit",
   "created_at": "2026-07-30T21:28:13.122Z",
-  "diffstat_sha256": "sha256:d608d79d0b109ef668fb62d9a08c58c2f2e150f124681825c8a38f8dc332607d",
+  "diffstat_sha256": "sha256:2ec2c7e15d7a16e16370822491c97c8afb1c92b1ff91eaa766f37272456707b9",
   "last_verified_at": "2026-07-30T22:05:59.806Z",
   "last_verified_diffstat_sha256": "sha256:c4b15a21ba94abca976037d786139e79110110fb214be9dbc0ec4ff391ee703c",
   "pr_number": 4696,
